hydropower/requirements.txt
Django==2.0.8
Markdown==2.6.11
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.8.2
jsonfield==2.0.2
psycopg2==2.7.5
Pillow==5.1.0

backcall==0.1.0   
decorator==4.3.0    
django-bootstrap3==10.0.1  
django-colorfield==0.1.15   
ipdb==0.11    
ipython==6.5.0   
ipython-genutils==0.2.0   
jedi==0.12.1   
numpy==1.15.0  
pandas==0.23.4  
parso==0.3.1   
pexpect==4.6.0   
pickleshare==0.7.4     
pip==18.0    
prompt-toolkit==1.0.15    
ptyprocess==0.6.0   
Pygments==2.2.0   
python-dateutil==2.7.3   
pytz==2018.5  
setuptools==40.1.0  
simplegeneric==0.8.1   
six==1.11.0  
sqlparse==0.2.4   
traitlets==4.3.2   
wcwidth==0.1.7   
xlrd==1.1.0 